NMS-16571 : added testcase in NodeRestServiceIT
